Organize directory into a project
