+++ TwoMQTT/Core/Models/MQTTDiscovery.cs
