+++ TwoMQTT/Core/Models/MQTTDiscoveryDevice.cs
