+++ TwoMQTT/Core/Models/SharedCommand.cs
